# laravel-test/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "RESTFul API",
        "description": "API to manage items and authentication",
        "version": "1.0"
    },
    "paths": {
        "/api/login": {
            "post": {
                "summary": "Login",
                "operationId": "a3b306d14572d1f4bd6c064b3233e7b8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "email": "user@example.com",
                                    "password": "password"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful login"
                    },
                    "401": {
                        "description": "Invalid Credentials"
                    }
                }
            }
        },
        "/api/items": {
            "get": {
                "summary": "List all items",
                "operationId": "2cbf07eb95d033f8775c36424ac8fa06",
                "responses": {
                    "200": {
                        "description": "List of items"
                    }
                }
            }
        },
        "/api/items/{id}": {
            "get": {
                "summary": "Buscar um item pelo ID",
                "operationId": "7f1ffbd9fa2b704c61ebcccb636ebb27",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Item Details"
                    },
                    "404": {
                        "description": "Item not found"
                    }
                }
            }
        }
    }
}